create separate file for tlsutils
